pyproject.toml
[tool.poetry]
name = "surv-backend"
version = "0.1.0"
description = "Maxwell's Magic Backend"
authors = ["bfpill <max@v3rv.com>"]
license = "proprietary"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9 <3.13"
blinker ="1.7.0"
click ="8.1.7"
Flask ="3.0.0"
Flask-Cors ="4.0.0"
gunicorn ="21.2.0"
itsdangerous ="2.1.2"
Jinja2 ="3.1.2"
MarkupSafe ="2.1.3"
packaging ="23.2"
python-dotenv ="1.0.0"
uWSGI ="2.0.23"
Werkzeug ="3.0.1"
fastapi = "^0.109.0"
pydantic-settings = "^2.1.0"
openai = "^1.7.2"
uvicorn = "^0.25.0"
rich = "^13.7.0"
orjson = "^3.9.12"
requests = "^2.31.0"
firebase-admin = "^6.4.0"
pinecone-client = "3.0.0"
scikit-learn = "^1.4.1.post1"
pillow = "^10.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
